requirements.txt
requests
pandas
jinja2
python-dotenv
PyYAML
openpyxl
fastapi
uvicorn
